folder batch support, add labelmap

+++ splitData.py
@@ -20,7 +20,8 @@ def splitData(
     dst_dir='./cache',
     ext='jpg',
     train_val_ratio=0.9,
-    seed=None
+    seed=None,
+    clean_dst=False
 ):
   ## 乱数設定
   if seed:
@@ -29,22 +30,25 @@ def splitData(
 
   ## train用とeval用に画像ファイルを分割する
   train_dir = os.path.join(dst_dir, 'train')
-  eval_dir = os.path.join(dst_dir, 'eval')
+  eval_dir = os.path.join(dst_dir, 'val')
   print('Splitting {} -> {}, {}...'.format(src_dir, train_dir, eval_dir))
 
   ## 入力ディレクトリが存在しない場合強制終了
   assert os.path.isdir(os.path.join(src_dir, 'Annotations')), 'Not found Annotations at {}'.format(src_dir)
-  assert os.path.isdir(os.path.join(src_dir, 'Images')), 'Not found Images at {}'.format(src_dir)
+  assert os.path.isdir(os.path.join(src_dir, ext+'Images')), 'Not found Images at {}'.format(src_dir)
 
   ## 出力ディレクトリの用意
-  if os.path.isdir(train_dir):
-    shutil.rmtree(train_dir)
-  os.makedirs(os.path.join(train_dir, "Annotations"))
-  os.makedirs(os.path.join(train_dir, "Images"))
-  if os.path.isdir(eval_dir):
-    shutil.rmtree(eval_dir)
-  os.makedirs(os.path.join(eval_dir, "Annotations"))
-  os.makedirs(os.path.join(eval_dir, "Images"))
+  if clean_dst:
+    if os.path.isdir(train_dir):
+      shutil.rmtree(train_dir)
+    if os.path.isdir(eval_dir):
+      shutil.rmtree(eval_dir)
+
+  os.makedirs(os.path.join(train_dir, "Annotations"), exist_ok=True)
+  os.makedirs(os.path.join(train_dir, ext+"Images"), exist_ok=True)
+
+  os.makedirs(os.path.join(eval_dir, "Annotations"), exist_ok=True)
+  os.makedirs(os.path.join(eval_dir, ext+"Images"), exist_ok=True)
 
   ## データ分割開始
   xml_files = glob(os.path.join(src_dir, "Annotations/*.xml"))
@@ -53,15 +57,18 @@ def splitData(
   def move_to(set_name, xml_paths):
     for xml_path in tqdm(xml_paths):
       file_name = os.path.splitext(os.path.basename(xml_path))[0]
-      jpeg_path = os.path.join(src_dir, "Images", "{}.{}".format(file_name, ext))
+      jpeg_path = os.path.join(src_dir, ext+"Images", "{}.{}".format(file_name, ext))
       dst_xml_path = os.path.join(dst_dir, set_name, "Annotations", "{}.xml".format(file_name))
-      dst_img_path = os.path.join(dst_dir, set_name, "Images", "{}.{}".format(file_name, ext))
-      shutil.copyfile(xml_path, dst_xml_path)
-      shutil.copyfile(jpeg_path, dst_img_path)
+      dst_img_path = os.path.join(dst_dir, set_name, ext+"Images", "{}.{}".format(file_name, ext))
+      if os.path.isfile(jpeg_path):
+        shutil.copyfile(xml_path, dst_xml_path)
+        shutil.copyfile(jpeg_path, dst_img_path)
+      else:
+        print('skip', jpeg_path)
 
   train_num = int(len(xml_files) * train_val_ratio)
   move_to("train", xml_files[:train_num])
-  move_to("eval", xml_files[train_num:])
+  move_to("val", xml_files[train_num:])
 
 if __name__ == "__main__":
   parser = argparse.ArgumentParser()
